brotli: use uint32 instead of uint8 and uint16
A 32-bit word can be loaded with a single instruction, while uint8 and uint16
require extra masking, costing us in CPU time. This slightly increase memory
utilization.

As an aside, wider words allow us to process prefix codes with more bits.
This makes bit_reader.go useful for bzip2, which can have 24-bit prefix codes.

benchmark                              old MB/s     new MB/s     speedup
BenchmarkDecodeDigitsSpeed1e4-4        54.15        58.89        1.09x
BenchmarkDecodeDigitsSpeed1e5-4        62.22        64.24        1.03x
BenchmarkDecodeDigitsSpeed1e6-4        65.07        66.47        1.02x
BenchmarkDecodeDigitsDefault1e4-4      48.42        49.30        1.02x
BenchmarkDecodeDigitsDefault1e5-4      61.66        63.62        1.03x
BenchmarkDecodeDigitsDefault1e6-4      67.63        68.38        1.01x
BenchmarkDecodeDigitsCompress1e4-4     49.02        50.12        1.02x
BenchmarkDecodeDigitsCompress1e5-4     61.32        62.89        1.03x
BenchmarkDecodeDigitsCompress1e6-4     66.14        66.08        1.00x
BenchmarkDecodeTwainSpeed1e4-4         55.66        56.40        1.01x
BenchmarkDecodeTwainSpeed1e5-4         72.98        73.83        1.01x
BenchmarkDecodeTwainSpeed1e6-4         79.83        81.02        1.01x
BenchmarkDecodeTwainDefault1e4-4       49.05        48.86        1.00x
BenchmarkDecodeTwainDefault1e5-4       73.79        74.80        1.01x
BenchmarkDecodeTwainDefault1e6-4       85.55        87.26        1.02x
BenchmarkDecodeTwainCompress1e4-4      49.84        49.41        0.99x
BenchmarkDecodeTwainCompress1e5-4      75.59        74.74        0.99x
BenchmarkDecodeTwainCompress1e6-4      85.44        86.81        1.02x

benchmark                              old bytes     new bytes     delta
BenchmarkDecodeDigitsSpeed1e4-4        28042         28760         +2.56%
BenchmarkDecodeDigitsSpeed1e5-4        97562         101946        +4.49%
BenchmarkDecodeDigitsSpeed1e6-4        98151         102999        +4.94%
BenchmarkDecodeDigitsDefault1e4-4      33705         39069         +15.91%
BenchmarkDecodeDigitsDefault1e5-4      108796        122607        +12.69%
BenchmarkDecodeDigitsDefault1e6-4      110062        124974        +13.55%
BenchmarkDecodeDigitsCompress1e4-4     32442         36713         +13.17%
BenchmarkDecodeDigitsCompress1e5-4     108460        121182        +11.73%
BenchmarkDecodeDigitsCompress1e6-4     110798        125630        +13.39%
BenchmarkDecodeTwainSpeed1e4-4         33403         38410         +14.99%
BenchmarkDecodeTwainSpeed1e5-4         100138        106204        +6.06%
BenchmarkDecodeTwainSpeed1e6-4         100551        106951        +6.36%
BenchmarkDecodeTwainDefault1e4-4       36978         43663         +18.08%
BenchmarkDecodeTwainDefault1e5-4       109996        124911        +13.56%
BenchmarkDecodeTwainDefault1e6-4       111502        127694        +14.52%
BenchmarkDecodeTwainCompress1e4-4      35848         41897         +16.87%
BenchmarkDecodeTwainCompress1e5-4      108956        123071        +12.95%
BenchmarkDecodeTwainCompress1e6-4      114126        132158        +15.80%

+++ brotli/bit_reader.go
@@ -188,7 +188,7 @@ func (br *bitReader) TryReadSymbol(pd *prefixDecoder) (uint, bool) {
 	if br.numBits < uint(pd.minBits) || len(pd.chunks) == 0 {
 		return 0, false
 	}
-	chunk := pd.chunks[uint16(br.bufBits)&pd.chunkMask]
+	chunk := pd.chunks[uint32(br.bufBits)&pd.chunkMask]
 	nb := uint(chunk & prefixCountMask)
 	if nb > br.numBits || nb > uint(pd.chunkBits) {
 		return 0, false
@@ -207,11 +207,11 @@ func (br *bitReader) ReadSymbol(pd *prefixDecoder) uint {
 	nb := uint(pd.minBits)
 	for {
 		br.FeedBits(nb)
-		chunk := pd.chunks[uint16(br.bufBits)&pd.chunkMask]
+		chunk := pd.chunks[uint32(br.bufBits)&pd.chunkMask]
 		nb = uint(chunk & prefixCountMask)
 		if nb > uint(pd.chunkBits) {
 			linkIdx := chunk >> prefixCountBits
-			chunk = pd.links[linkIdx][uint16(br.bufBits>>pd.chunkBits)&pd.linkMask]
+			chunk = pd.links[linkIdx][uint32(br.bufBits>>pd.chunkBits)&pd.linkMask]
 			nb = uint(chunk & prefixCountMask)
 		}
 		if nb <= br.numBits {
@@ -246,14 +246,14 @@ func (br *bitReader) ReadPrefixCode(pd *prefixDecoder, maxSyms uint) {
 func (br *bitReader) readSimplePrefixCode(pd *prefixDecoder, maxSyms uint) {
 	var codes [4]prefixCode
 	nsym := int(br.ReadBits(2)) + 1
-	clen := neededBits(uint16(maxSyms))
+	clen := neededBits(uint32(maxSyms))
 	for i := 0; i < nsym; i++ {
-		codes[i].sym = uint16(br.ReadBits(clen))
+		codes[i].sym = uint32(br.ReadBits(clen))
 	}
 
 	var copyLens = func(lens []uint) {
 		for i := 0; i < nsym; i++ {
-			codes[i].len = uint8(lens[i])
+			codes[i].len = uint32(lens[i])
 		}
 	}
 	var compareSwap = func(i, j int) {
@@ -299,7 +299,7 @@ func (br *bitReader) readComplexPrefixCode(pd *prefixDecoder, maxSyms, hskip uin
 	for _, sym := range complexLens[hskip:] {
 		clen := br.ReadSymbol(&decCLens)
 		if clen > 0 {
-			codeCLensArr[sym] = prefixCode{sym: uint16(sym), len: uint8(clen)}
+			codeCLensArr[sym] = prefixCode{sym: uint32(sym), len: uint32(clen)}
 			if sum -= 32 >> clen; sum <= 0 {
 				break
 			}
@@ -325,7 +325,7 @@ func (br *bitReader) readComplexPrefixCode(pd *prefixDecoder, maxSyms, hskip uin
 		if clen < 16 {
 			// Literal bit-length symbol used.
 			if clen > 0 {
-				codes = append(codes, prefixCode{sym: uint16(sym), len: uint8(clen)})
+				codes = append(codes, prefixCode{sym: uint32(sym), len: uint32(clen)})
 				clenLast = clen
 				sum -= 32768 >> clen
 			}
@@ -353,7 +353,7 @@ func (br *bitReader) readComplexPrefixCode(pd *prefixDecoder, maxSyms, hskip uin
 			if repSym == 16 {
 				clen := clenLast
 				for symEnd := sym + repDiff; sym < symEnd; sym++ {
-					codes = append(codes, prefixCode{sym: uint16(sym), len: uint8(clen)})
+					codes = append(codes, prefixCode{sym: uint32(sym), len: uint32(clen)})
 				}
 				sum -= int(repDiff) * (32768 >> clen)
 			} else {
